integrate seriliaztion/deserialization to executable
