+++ python/tvm/relay/__init__.py
@@ -37,8 +37,6 @@ from . import param_dict
 from . import feature
 from .backend import vm
 from .backend import profiler_vm
-from .backend import serializer
-from .backend import deserializer
 from .backend import vmobj
 
 # Root operators
